requirements.txt
websocket-client
pandas
win10toast
pywin32
requests
